requirements.txt
numpy==1.19.2
torchvision==0.5.0
imageio==2.9.0
matplotlib==3.3.2
torch==1.4.0
opencv_python==4.4.0.46
tqdm==4.50.2
scipy==1.5.3
h5py==2.10.0
Pillow==8.2.0